update EnergyCostsBarChart component to use EnergyCostsBarChartCard for better code organization
